Resolve email domain as unambiguous fully qualified domain name (#30)
If dnspython receives a domain name that does not end in a period it
may also try resolving it as not fully qualified. This can cause
domains to be incorrectly resolved as if they exist.

This can be tested by adding a line `search mailinator.com` to
/etc/resolv.conf. With that line added, the address `test@asdf.asdf`
will be considered deliverable by email-validator.

The solution is to append a period to the domain before passing it to
dnspython. This is valid because the domain part of the email address
is not allowed to end in a period, and is required to be interpreted
as fully qualified.