Fix exception on DNS timeout
In 8243bd238fbfbbef9522471ded214420e1b1fe7e, I re-organized the return value of validate_email to be an object rather than a dict. However I didn't properly handle the case when there's a DNS timeout.

I think my original intention was to hide timeout errors because it's unclear if a timeout is a local problem or a problem with the email's domain name. If it's the former, passing through the timeout exception to the caller would make sense. But if it's something that can be triggered by user input, then it shouldn't raise a timeout exception, it should raise a EmailUndeliverableError. Since we can't decide which is the case, we treat it as if the deliverability check was not performed at all.

Fixes #41