Handle NoNameservers and NXDOMAIN exceptions better
I am not sure what NoNameservers means, so I think it might be that no local nameservers could respond. Local error conditions should not fail deliverability (same as a timeout).

NXDOMAIN means no records are present for a domain, so a NXDOMAIN after the MX query can skip straight to rejecting the domain, rather than going to the A/AAAA fallback. Consequently, the error for a missing fallback can be changed from "does not exist" (which is now handled by the NXDOMAIN except block) to "does not accept email."