Fix the domain name length limit
I previously copied the domain name length limit from the RFCs, but I misunderstood that "octets" in the RFCs didn't mean the number of characters in the ASCII domain name but the number of bytes as transmitted. When transmitted, the domain name has one byte for each label (part between periods) giving the label length. Those bytes correspond to the dots, except for the last label which doesn't have a dot, and the empty label which isn't printed. So the longest domain name length in characters is two less than what I thought.

See https://stackoverflow.com/questions/32290167/what-is-the-maximum-length-of-a-dns-name for explanation.

I noticed this when I saw that the idna package was rejecting domain names with 254 characters which this library accepted.